audit(2.9.7): internal audit sections A-I filled in + bugfixes
INTERNAL_AUDIT.md:
- All section tables A-01 through I-06 filled with real results
- Findings Log written: F-001/F-003 resolved (false positives),
  F-002 FIXED, F-004 pending (SSH server), F-005 FIXED (compile)
- Overview table: 0 critical open bugs remaining
- Critical blockers checklist updated: C-02a ✅, H-01c ✅, A-02a ✅

Bugfixes:
- fix(core): verify_chain() .unwrap() → .ok_or_else() in chain.rs:428
  Prevents panic on chain gap during disk corruption (F-002)
- fix(pool): wallet.rs u64→u128 type mismatch (3 locations) in
  payout/wallet.rs — amount_atomic cast, sum type, Recipient (F-005)

False positives cleared:
- F-001: Share replay protection EXISTS in validator.rs:128 via
  cache HashMap<String, CacheEntry> with TTL=600s
- F-003: All docs correctly state 144,000,000,000 ZION (144B)

cargo check: 0 errors ✅

+++ L1/pool/src/payout/wallet.rs
@@ -92,7 +92,7 @@ impl PoolWallet {
                     .and_then(|v| v.as_str())
                     .unwrap_or("")
                     .to_string();
-                let amount = u.get("amount_atomic").and_then(|v| v.as_u64()).unwrap_or(0);
+                let amount = u.get("amount_atomic").and_then(|v| v.as_u64()).unwrap_or(0) as u128;
                 let addr = u
                     .get("address")
                     .and_then(|v| v.as_str())
@@ -120,7 +120,7 @@ impl PoolWallet {
         tracing::info!(
             "💰 Pool wallet UTXOs: {} found, total balance = {} ZION",
             all_utxos.len(),
-            all_utxos.iter().map(|u| u.amount).sum::<u64>() as f64 / 1_000_000.0
+            all_utxos.iter().map(|u| u.amount).sum::<u128>() as f64 / 1_000_000.0
         );
 
         Ok(all_utxos)
@@ -202,7 +202,7 @@ impl PoolWallet {
     pub async fn send_single_payout(&self, to_address: &str, amount_atomic: u64) -> Result<String> {
         let recipients = vec![Recipient {
             address: to_address.to_string(),
-            amount: amount_atomic,
+            amount: amount_atomic as u128,
         }];
 
         let result = self.send_batch_payout(recipients).await?;
